优化，哥们！优化！修复bug，哥们！修复bug！

+++ internal/model/notification.go
@@ -17,7 +17,7 @@ type GetNotificationsInput struct {
 }
 
 type NotificationBase struct {
-	Id              int    `json:"id"         orm:"id"          dc:"提醒ID"`
+	Id              int64  `json:"id"         orm:"id"          dc:"提醒ID"`
 	QuestionId      int    `json:"question_id" orm:"question_id" dc:"问题ID"`
 	QuestionTitle   string `json:"question_title"                dc:"问题标题"`
 	QuestionContent string `json:"question_content"              dc:"问题内容"`
